Gerber and Drill files
Here are the .grb files for getting the board printed. It's just through hole so you'll need to solder the parts onto it. 7 x 220 resistors, 1 tiltball switch (at the top 2 pin through hole, 6 buttons, cr2450 battery and battery holder, and of course and atmega328p. You have to flash the 8 megahertz bootloader onto the chip so it'll run without an external crystal. There are videos on youtube. Use the latest version of the code, v4 in main4 branch.